+++ src/api.cpp
@@ -1012,14 +1012,14 @@ void rgl::rgl_material(int *successptr, int* idata, char** cdata, double* ddata)
 
   mat.alphablend  = false;
   
-  size_t len_user_id = strlen(cdata[0]);
-  if (len_user_id) {
-    char* in_user_id = new char [len_user_id + 1];
-    strncpy(in_user_id, cdata[0], len_user_id);
-    in_user_id[len_user_id] = '\0';
-    mat.user_id = string(in_user_id);
+  size_t len_tag = strlen(cdata[0]);
+  if (len_tag) {
+    char* in_tag = new char [len_tag + 1];
+    strncpy(in_tag, cdata[0], len_tag);
+    in_tag[len_tag] = '\0';
+    mat.tag = string(in_tag);
   } else
-    mat.user_id = string();
+    mat.tag = string();
   
   if ( nfilenames > 0 ) {
     mat.texture = new Texture(nfilenames, pixmapfn, mat.textype, mat.mipmap, mat.minfilter, mat.magfilter, mat.envmap);
@@ -1137,11 +1137,11 @@ void rgl::rgl_getmaterial(int *successptr, int *id, int* idata, char** cdata, do
   } else 
     idata[10] = 0;
   
-  /* Can't use user_id.length() here, because R has highjacked length() */
-  size_t len_user_id = strlen(mat->user_id.c_str());
-  cdata[0] = R_alloc(len_user_id + 1, 1);
-  strncpy(cdata[0], mat->user_id.c_str(), len_user_id);
-  (cdata[0])[len_user_id] = '\0';
+  /* Can't use tag.length() here, because R has highjacked length() */
+  size_t len_tag = strlen(mat->tag.c_str());
+  cdata[0] = R_alloc(len_tag + 1, 1);
+  strncpy(cdata[0], mat->tag.c_str(), len_tag);
+  (cdata[0])[len_tag] = '\0';
   
   CHECKGLERROR;