+++ NEWS.md
@@ -8,7 +8,7 @@ will be used as a "mipmap".  This may be used in
 of the display.
 * Added `getBoundary3d()` function to extract the boundary
 edges of a mesh.
-* Added material property `user_id`, a string associated 
+* Added material property `tag`, a string associated 
 with each object.  The value is reported by `ids3d()` and
 may be used to select objects in `pop3d()`, but otherwise is 
 largely ignored by `rgl`.